% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringSomas.R
\name{clusteringSomas}
\alias{clusteringSomas}
\title{directional_parents is a boolean value denoting if the linear variables should have directional parents}
\usage{
clusteringSomas(data, is_dir, num_clusters = 10, num_parents = 5,
  directional_parents = T, cores = 1)
}
\description{
directional_parents is a boolean value denoting if the linear variables should have directional parents
}
\examples{
model <- clusteringSomas(dataset$data, dataset$is_dir, num_clusters = 3, cores = 10)
}
