% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_KL.R
\name{compute_directional_KL}
\alias{compute_directional_KL}
\title{Compute KL divergence of all directional variables}
\usage{
compute_directional_KL(model_p, model_q, cluster_p, cluster_q)
}
\description{
Compute KL divergence of all directional variables
}
\examples{
vM_p <- model
vM_q <- model
cluster_p <- 1
cluster_q <- 2
compute_directional_KL(vM_p,vM_q)
}
