% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_KL.R
\name{compute_linear_KL}
\alias{compute_linear_KL}
\title{Compute KL divergence of linear variables}
\usage{
compute_linear_KL(model_p, model_q, cluster_p, cluster_q, is_dir)
}
\description{
Compute KL divergence of linear variables
}
\examples{
cluster_p <- 2
cluster_q <- 2
is_dir <- dataset$is_dir
compute_linear_KL(model, clustering_model, cluster_p, cluster_q, is_dir)
}
