% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizationParams.R
\name{expectation}
\alias{expectation}
\title{Compute E-step}
\usage{
expectation(priori, gaussParams, vMParams, linearData, angularData)
}
\arguments{
\item{priori}{is a vector that represents the a priori probability of an instance belonging to an instace without further information}

\item{gaussParams}{is a bn.fit object that saves the parameters of the multivariate gaussian Bayesian network}

\item{vMParams}{is a list saving objects of vonMises Class for each cluster with their parameters mu and kappa}

\item{linearData}{a dataframe of the variables that represents linear variables}

\item{angularData}{a dataframe of the variables that represents angular variables}
}
\value{
list with two entries, the first one is the matrix of responsabilities the second one is de log-likelihood of the model
}
\description{
Compute responsabilities of each cluster on each instance
}
