% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_artificial_BN.R
\name{generate_artificial_BN}
\alias{generate_artificial_BN}
\title{Adaptation of pcalg method to generate artificial Gaussian BNs}
\usage{
generate_artificial_BN(is_dir, prob = 0.3, lB = -10, uB = 10,
  node_names = as.character(1:n), num_clusters = 3)
}
\arguments{
\item{is_dir}{is a binary array 1xn where n is the number of variables, it takes value 1 when the variable is directional and 0 if the variable is linear}

\item{prob}{probability to add an edge between two variables}

\item{lB}{lower bound for the coefficients}

\item{uB}{upper bound for the coefficients}

\item{node_names}{an array with the names of the nodes}

\item{num_clusters}{is an integer denoting the number of clusters}
}
\value{
model is a list generated by the functions generate_artificial_BN or clusteringSomas
}
\description{
Adaptation of pcalg method to generate artificial Gaussian BNs
}
\examples{
is_dir <- c(rep(1,3),rep(0,7))
model  <- generate_artificial_BN(is_dir)

}
