% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insertion_point.R
\name{get_insertion_point}
\alias{get_insertion_point}
\title{Get insertion point of somas}
\usage{
get_insertion_point(path_complete_neuron, path_soma, path_off_export,
  num_curves)
}
\arguments{
\item{directory}{path to the folder where PLY files of the somas are saved}

\item{output_shape_diameter}{path to the folder where somas with shape diameter will be saved}

\item{broken_mesh}{path to the folder where somas will be saved after that the GMM threshold removes the dendrites}

\item{output_poisson_reconstruction}{path to the folder where somas will be saved after the mesh closing process}

\item{final_result}{path to the folder where final segmented somas will be saved after the isolated pieces are removed}
}
\value{
None
}
\description{
Apply Gaussian mixture models to segment soma from dendrites
}
\examples{
path <- "/home/universidad/RStudio Projects/SomaSimulation/data"
soma_parameters <- get_insertion_point(path_complete_neuron = file.path(path,"ao_results"), path_soma = file.path(path, "sdf_results"), path_off_export = file.path(tempdir(),"soma_off_files"), num_curves = 6)

}
