% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_artificial_BN.R
\name{sampleBN}
\alias{sampleBN}
\title{Function to sample instances from an artificial BN}
\usage{
sampleBN(model, is_dir, cluster_id, num_simulations)
}
\arguments{
\item{model}{is a list generated by the functions generate_artificial_BN or clusteringSomas}

\item{is_dir}{is a binary array 1xn where n is the number of variables, it takes value 1 when the variable is directional and 0 if the variable is linear}

\item{cluster_id}{the id of the cluster from the instance should be sample}

\item{num_simulations}{number of instances to simulate from}
}
\value{
a matrix with the data simulated
}
\description{
Function to sample instances from an artificial BN
}
\examples{
is_dir <- c(rep(1,3),rep(0,7))
model  <- generate_artificial_BN(is_dir)
data <- sampleBN(model, is_dir, 1, 100)

}
