% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSoma.R
\name{simulation3Dmesh}
\alias{simulation3Dmesh}
\title{3D virtual spine}
\usage{
simulation3Dmesh(newSomas, idx, iterations = 4)
}
\arguments{
\item{newSomas}{a matrix where each row is a virtual spine}

\item{idx}{is number of the row number of the spine to representate}
}
\value{
rgl object 3D mesh3d
}
\description{
Generate the 3D representation of the virtual spines
}
\examples{
newSomas <- sampleSomas(bestModel, dataset$is_dir, cluster_id = 1, num_simulations = 20)
mesh <- simulation3Dmesh(newSomas,1,iterations=0)
shade3d(mesh,col="red")
mesh <- simulation3Dmesh(newSomas,2,iterations=0)
shade3d(mesh,col="red")
mesh <- simulation3Dmesh(newSomas,7,iterations=0)
shade3d(mesh,col="red")



}
