% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_KL.R
\name{vM_KL}
\alias{vM_KL}
\title{Compute KL divergence between two von Mises distributions}
\usage{
vM_KL(vM_p_var, vM_q_var)
}
\description{
Compute KL divergence between two von Mises distributions
}
\examples{
vM_p_var <- model$vMParams[[1]]$theta_2
vM_q_var <- model$vMParams[[2]]$theta_2
}
